Move ListTileTheme and its tests into separate classes and  add visualDensity to the ListTileTheme
