+++ packages/flutter/lib/src/material/checkbox_list_tile.dart
@@ -6,6 +6,7 @@ import 'package:flutter/widgets.dart';
 
 import 'checkbox.dart';
 import 'list_tile.dart';
+import 'list_tile_theme.dart';
 import 'theme.dart';
 import 'theme_data.dart';
