+++ packages/flutter/lib/src/material/drawer.dart
@@ -9,6 +9,7 @@ import 'colors.dart';
 import 'debug.dart';
 import 'drawer_theme.dart';
 import 'list_tile.dart';
+import 'list_tile_theme.dart';
 import 'material.dart';
 import 'material_localizations.dart';
 import 'theme.dart';
